Add files via upload
In the game, the player controls Pacman, a character represented by a yellow, circular figure with an ever-hungry mouth. The primary objective is to guide Pacman through a maze filled with small dots (pellets). Pacman must eat all the pellets in the maze while avoiding the four roaming ghosts — each programmed with distinct movement patterns and strategies to catch the player.

If a ghost catches Pacman, the player loses a life. The game typically ends when all lives are lost or when the player successfully consumes all the pellets in the maze. The player earns points by eating pellets, with bonus points awarded for eating power pellets (larger dots), which temporarily turn the ghosts blue and vulnerable. During this phase, Pacman can chase and eat the ghosts, sending them back to their starting position.